+++ src/commands/registerCommands.ts
@@ -5,9 +5,10 @@
 
 import { commands } from "vscode";
 import { IActionContext, registerCommand as registerCommandAzUI } from "vscode-azureextensionui";
-import { configure, configureApi } from "../configureWorkspace/configure";
-import { configureCompose } from "../configureWorkspace/configureCompose";
 import { ext } from "../extensionVariables";
+import { scaffold } from "../scaffolding/scaffold";
+import { scaffoldCompose } from "../scaffolding/scaffoldCompose";
+import { scaffoldDebugConfig } from "../scaffolding/scaffoldDebugConfig";
 import { deployImageToAzure } from "../utils/lazyLoad";
 import { viewAzureTaskLogs } from "../utils/lazyLoad";
 import { composeDown, composeRestart, composeUp } from "./compose";
@@ -51,6 +52,7 @@ import { registerWorkspaceCommand } from "./registerWorkspaceCommand";
 import { createAzureRegistry } from "./registries/azure/createAzureRegistry";
 import { deleteAzureRegistry } from "./registries/azure/deleteAzureRegistry";
 import { deleteAzureRepository } from "./registries/azure/deleteAzureRepository";
+import { deployImageToAci } from "./registries/azure/deployImageToAci";
 import { openInAzurePortal } from "./registries/azure/openInAzurePortal";
 import { buildImageInAzure } from "./registries/azure/tasks/buildImageInAzure";
 import { runAzureTask } from "./registries/azure/tasks/runAzureTask";
@@ -89,12 +91,13 @@ export function registerCommand(commandId: string, callback: (context: IActionCo
 }
 
 export function registerCommands(): void {
-    registerWorkspaceCommand('vscode-docker.api.configure', configureApi);
+    registerWorkspaceCommand('vscode-docker.configure', scaffold);
+    registerWorkspaceCommand('vscode-docker.configureCompose', scaffoldCompose);
+    registerWorkspaceCommand('vscode-docker.debugging.initializeForDebugging', scaffoldDebugConfig);
+
     registerWorkspaceCommand('vscode-docker.compose.down', composeDown);
     registerWorkspaceCommand('vscode-docker.compose.restart', composeRestart);
     registerWorkspaceCommand('vscode-docker.compose.up', composeUp);
-    registerWorkspaceCommand('vscode-docker.configure', configure);
-    registerWorkspaceCommand('vscode-docker.configureCompose', configureCompose);
     registerCommand('vscode-docker.pruneSystem', pruneSystem);
 
     registerWorkspaceCommand('vscode-docker.containers.attachShell', attachShellContainer);
@@ -132,6 +135,7 @@ export function registerCommands(): void {
     registerCommand('vscode-docker.registries.copyImageDigest', copyRemoteImageDigest);
     registerCommand('vscode-docker.registries.deleteImage', deleteRemoteImage);
     registerCommand('vscode-docker.registries.deployImageToAzure', deployImageToAzure);
+    registerCommand('vscode-docker.registries.deployImageToAci', deployImageToAci);
     registerCommand('vscode-docker.registries.disconnectRegistry', disconnectRegistry);
     registerCommand('vscode-docker.registries.help', registryHelp);
     registerWorkspaceCommand('vscode-docker.registries.logInToDockerCli', logInToDockerCli);